refactor: 위치정리
