+++ src/app/api/lounge/posts/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import type { GetPostsResponse, LoungeSortBy, SortOrder } from "@/types";
-import { getLoungePostsPageBFF } from "@/app/api/_services/lounge";
+import { getLoungePostsPageBFF } from "@/bff/lounge";
 
 export const dynamic = "force-dynamic";
