+++ src/app/lounge/page.tsx
@@ -4,8 +4,8 @@ import PrefetchBoundary from "@/components/boundary/PrefetchBoundary";
 import { Suspense } from "react";
 import LoungeSkeleton from "@/components/skeleton/LoungeSkeleton";
 import HotPostListSkeleton from "@/components/skeleton/HotPostListSkeleton";
-import { getLoungePostsPageBFF } from "@/app/api/_services/lounge";
-import { getHotPostsBFF } from "@/app/api/_services/hot";
+import { getLoungePostsPageBFF } from "@/bff/lounge";
+import { getHotPostsBFF } from "@/bff/hot";
 import { getNextPageParam } from "@/lib/pagination";
 import { QUERY_KEYS } from "@/constants/queryKey";
 import LoungeSearchSection from "./_components/LoungeSearchSection";
